third_party/WebKit/LayoutTests/inspector/elements/styles-4/styles-update-links-4.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script src="./styles-update-links.js"></script>
<style>
#pseudo::after {
    pseudo-property: "12";
    color: red;
}

#pseudo::after {
    border: 1px solid black;
}

#pseudo::before {
    color: blue;
}
</style>
<script>

function test()
{
    TestRunner.runTestSuite([
        function selectNodeWithPseudoElements(next)
        {
            ElementsTestRunner.selectNodeAndWaitForStyles("pseudo", next);
        },

        function testPseudoSectionPropertyEdit(next)
        {
            var treeItem = ElementsTestRunner.getMatchedStylePropertyTreeItem("pseudo-property");
            var treeElement = treeItem.section().addNewBlankProperty(1);
            ElementsTestRunner.waitForStyleApplied(onPropertyInserted);
            treeElement.applyStyleText("PROPERTY: INSERTED;", true);

            function onPropertyInserted()
            {
                TestRunner.addResult("\n\n#### AFTER PROPERTY INSERTED ####\n\n");
                ElementsTestRunner.dumpSelectedElementStyles(true, false, true);
                var rules = ElementsTestRunner.getMatchedRules();
                ElementsTestRunner.validateRuleRanges("pseudo", rules, next);
            }
        }
    ]);
}
</script>
<link rel="stylesheet" href="../styles/resources/styles-update-links-2.css"></link>
<link rel="stylesheet" href="../styles/resources/styles-update-links.css"></link>
</head>

<body onload="runTest()">
<p>
Tests that links are updated property when editing pseudo element property.
</p>

<div id="container" class="left-intact should-change">
Red text here.
</div>

<div id="other"></div>

<section id="pseudo"></div>

</body>
</html>
